Project owner can only see credentials action bar

+++ app/views/projects/partials/credentials.blade.php
@@ -11,20 +11,24 @@
 	     			<p><strong>Username:</strong> {{ $credential->username }}</p>
 	     			<p><strong>Password:</strong> {{ $credential->password }}</p>
 	     			<p><strong>Port:</strong> {{ $credential->port }}</p>
-	     			<div class="actions">
-	     				<ul class="inline-list list-style-none">
-	     					<li>
-		     				{{ Form::open(array('action' => 'CredentialsController@destroy', 'method' => 'delete')) }}
-		     					<input type="hidden" name="id" value="{{ $credential->id }}">
-		     					<button title="delete" id="{{ $credential->id }}" type="submit" class="btn btn-default btn-delete"><i class="fa fa-trash"></i></button>
-		     				{{ Form::close() }}							     						
-	     					</li>
-	     					<?php /*<li>
-	     						<button title="edit" class="btn btn-default"><a href=""><i class="fa fa-pencil"></i></a></button>				
-	     					</li>*/ ?>
-	     					<div class="clearfix"></div>
-	     				</ul>
-	     			</div>
+
+					@if($owner_id == Auth::id())
+						<div class="actions">
+							<ul class="inline-list list-style-none">
+								<li>
+								{{ Form::open(array('action' => 'CredentialsController@destroy', 'method' => 'delete')) }}
+									<input type="hidden" name="id" value="{{ $credential->id }}">
+									<button title="delete" id="{{ $credential->id }}" type="submit" class="btn btn-default btn-delete"><i class="fa fa-trash"></i></button>
+								{{ Form::close() }}
+								</li>
+								<?php /*<li>
+									<button title="edit" class="btn btn-default"><a href=""><i class="fa fa-pencil"></i></a></button>
+								</li>*/ ?>
+								<div class="clearfix"></div>
+							</ul>
+						</div>
+					@endif
+
 	     		</div>
      		@endif
      	@endforeach
@@ -32,7 +36,7 @@
   </div>	         	
 </div>
 
-	<!-- other panel -->
+<!-- other panel -->
 <div class="panel panel-default">
   <div class="panel-heading">Other Credentials</div>
   <div class="panel-body other--panel-body">					    
@@ -43,20 +47,21 @@
 	     			<h4>{{ $credential->name }}</h4	>
 	     			<p><strong>Username:</strong> {{ $credential->username }}</p>
 	     			<p><strong>Password:</strong> {{ $credential->password }}</p>
-	     			<div class="actions">
-	     				<ul class="inline-list list-style-none">
-	     					<li>
-		     				{{ Form::open(array('action' => 'CredentialsController@destroy', 'method' => 'delete')) }}
-		     					<input type="hidden" name="id" value="{{ $credential->id }}">
-		     					<button title="delete" id="{{ $credential->id }}" type="submit" class="btn btn-default btn-delete"><i class="fa fa-trash"></i></button>
-		     				{{ Form::close() }}							     						
-	     					</li>
-	     					<?php /*<li>
-	     						<button title="edit" class="btn btn-default"><a href=""><i class="fa fa-pencil"></i></a></button>				
-	     					</li>*/?>
-	     					<div class="clearfix"></div>
-	     				</ul>
-	     			</div>
+
+					@if($owner_id == Auth::id())
+						<div class="actions">
+							<ul class="inline-list list-style-none">
+								<li>
+								{{ Form::open(array('action' => 'CredentialsController@destroy', 'method' => 'delete')) }}
+									<input type="hidden" name="id" value="{{ $credential->id }}">
+									<button title="delete" id="{{ $credential->id }}" type="submit" class="btn btn-default btn-delete"><i class="fa fa-trash"></i></button>
+								{{ Form::close() }}
+								</li>
+								<div class="clearfix"></div>
+							</ul>
+						</div>
+					@endif
+
 	     		</div>
      		@endif
      	@endforeach
